+++ src/WorkspaceException.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Tivins\LlmBasic;
+
+class WorkspaceException extends \RuntimeException
+{
+}